Replace minunit.h with greatest.h
The minunit.h header file was used in order to quickly setup unit
tests for this project. With the unit tests starting to come together
I found it necessary to move to something more robust. The greatest.h
header seems to be the sweet spot between usability and robustness
so it will be used for the time being.